+++ src/Hacks/Speedhack.cpp
@@ -5,17 +5,36 @@
 using namespace geode::prelude;
 
 FMOD::ChannelGroup* masterGroup;
-
+FMOD::DSP *pitchDSP = nullptr;
+InputModule* pitchAmount;
+float lastPitch = 0;
 
 float speedhackLogic(float dt)
 {
     #ifndef GEODE_IS_IOS
     if (!masterGroup)
+    {
         FMODAudioEngine::sharedEngine()->m_system->getMasterChannelGroup(&masterGroup);
+    }
     #endif
 
-    if (!masterGroup)
+    if (!pitchAmount)
+        pitchAmount = as<InputModule*>(Client::GetModule("pitch-shifter")->options[0]);
+
+    if (!masterGroup || !pitchAmount)
         return dt;
+
+    if (lastPitch != (Client::GetModuleEnabled("pitch-shifter") ? pitchAmount->getFloatValue() : 1))
+    {
+        if (pitchDSP)
+            masterGroup->removeDSP(pitchDSP);
+
+        FMODAudioEngine::sharedEngine()->m_system->createDSPByType(FMOD_DSP_TYPE_PITCHSHIFT, &pitchDSP);
+        masterGroup->addDSP(0, pitchDSP);
+        lastPitch = Client::GetModuleEnabled("pitch-shifter") ? pitchAmount->getFloatValue() : 1;
+
+        pitchDSP->setParameterFloat(FMOD_DSP_PITCHSHIFT_PITCH, Client::GetModuleEnabled("pitch-shifter") ? pitchAmount->getFloatValue() : 1);
+    }
 
     ColourUtility::totalSessionTime += dt;
 
@@ -23,22 +42,9 @@ float speedhackLogic(float dt)
     {
         if (SpeedhackEnabled::instance->enabled)
         {
-            float v = 1.0f;
-
-            auto x = numFromString<float>(SpeedhackTop::instance->text);
-
-            if (x.isOk())
-            {
-                v = x.value();
-            }
-
-            if (v < 0.01f)
-                v = 0.01f;
-
-            if (v > 99999)
-                v = 99999;
-
             bool m = SpeedhackMus::instance->enabled;
+
+            float v = SpeedhackTop::instance->getFloatValue();
 
             if (SpeedhackGameplay::instance->enabled)
                 if (!(PlayLayer::get() || GameManager::sharedState()->getEditorLayer())) { v = 1.0f; }
@@ -75,10 +81,13 @@ class $modify (CCScheduler)
 //
 
 FMOD_RESULT MenuLayer_onNewgrounds(FMOD::System* self, const char *name, FMOD::ChannelGroup **channelgroup) {
-    auto res = self->createChannelGroup(name, channelgroup);
+    //auto res = self->createChannelGroup(name, channelgroup);
+    auto res = reinterpret_cast<FMOD_RESULT(__thiscall*)(FMOD::System*, const char*, FMOD::ChannelGroup**)>(geode::base::get() + 0x4c8964)(self, name, channelgroup);
 
     if (!masterGroup)
         masterGroup = *channelgroup;
+
+    log::info("WE HAVE A MASTER GROUP LETS FUCKING GO!!!!");
 
     return res;
 }
@@ -87,7 +96,7 @@ $execute {
     Mod::get()->hook(
         reinterpret_cast<void*>(geode::base::get() + 0x4c8964), // address
         &MenuLayer_onNewgrounds, // detour
-        "MenuLayer::onNewgrounds", // display name, shows up on the console
+        "FMOD::System::createChannelGroup", // display name, shows up on the console
         tulip::hook::TulipConvention::Thiscall // calling convention
     );
 }